refactor: relocate modules
